socketserver/requirements.txt
ngo==1.9
backports.ssl-match-hostname==3.4.0.2
django-nose==1.4.1
django-redis-sessions==0.5.0
django-websocket-redis==0.4.5
gevent==1.0.2
greenlet==0.4.9
nose==1.3.7
redis==2.10.3
requests==2.7.0
six==1.10.0
websocket-client==0.35.0
